packages/monitorme-server-agent-py/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monitorme-server-agent-py"
version = "0.1.0"
description = "MonitorMe Python server agent for tracing and baggage propagation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "opentelemetry-api>=1.24.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-zipkin-json>=1.24.0",
  "opentelemetry-instrumentation-fastapi>=0.45b0",
  "opentelemetry-instrumentation-requests>=0.45b0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["monitorme_server_agent_py*"]
